+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+### [[Unreleased](https://github.com/Zeroji/semicolon/releases/latest)]
+
+#### Added
+
++ MISC (Installs Semicolon Cogs) configuration / installation tool
+
 ### [[v0.2.1](https://github.com/Zeroji/semicolon/releases/tag/v0.2.1)]
 
 #### Added